api/pyproject.toml
[project]
name = "ai-podcast"
version = "0.1.0"
description = ""
authors = [
    {name = "Sairam M",email = "sairam.m.2005.65@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "gtts (>=2.5.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "google-cloud-texttospeech (>=2.25.1,<3.0.0)",
    "google-genai (>=1.10.0,<2.0.0)",
    "fuzzywuzzy (>=0.18.0,<0.19.0)",
    "python-levenshtein (>=0.27.1,<0.28.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "alembic (>=1.16.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "inngest (>=0.4.22,<0.5.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
